:truck: move user trainings to `/src/devcomp`
